v2/requirements.txt
beautifulsoup4==4.11.1
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.6
cryptography @ file:///C:/b/abs_36x9ifdcl4/croot/cryptography_1665612655344/work
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.1
httplib2==0.21.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
joblib==1.2.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
nltk==3.8.1
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
oauthlib==3.2.2
Pillow==9.3.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605307512533/work
regex==2022.10.31
requests @ file:///C:/ci/requests_1657735342357/work
requests-oauthlib==1.3.1
rsa==4.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.3.2.post1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_21qa0j12xt/croot/urllib3_1670527000672/work
wikipedia==1.4.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
